+++ src/lib/components/ui/spinner/spinner.svelte
@@ -1,23 +1,17 @@
 <script lang="ts">
   import { cn } from "$lib/utils.js";
-  import Loader2Icon from "@lucide/svelte/icons/loader-2";
-  import type { SVGAttributes } from "svelte/elements";
+  import { Loader2, type LucideProps } from "@lucide/svelte";
 
   let {
     class: className,
     role = "status",
-    // we add color and stroke for compatibility with different icon libraries props
-    color,
-    stroke,
     "aria-label": ariaLabel = "Loading",
     ...restProps
-  }: SVGAttributes<SVGSVGElement> = $props();
+  }: Partial<LucideProps> = $props();
 </script>
 
-<Loader2Icon
+<Loader2
   {role}
-  color={color === null ? undefined : color}
-  stroke={stroke === null ? undefined : stroke}
   aria-label={ariaLabel}
   class={cn("size-4 animate-spin", className)}
   {...restProps}